feat: check syntax in ci
