+++ .github/workflows/ci.yaml
@@ -22,5 +22,7 @@ jobs:
           cache: 'pnpm'
       - name: Install Dependencies
         run: pnpm install
+      - name: Check
+        run: npm run check
       - name: Build
         run: npm run build
